Improve package discovery: add package `Bundle`s
Bundles are a way to group a bunch of package classes, and then `enable`
them on-demand. Enabling causes a package class to be instantiated; then
`inject` is sent to the package instance and package vocabulary is
injected into the so-called bundle block, which is owned by the bundle
but accessible from the outside.

Overall package discovery didn't change though, but now at least you
have a choice to not use it and spin up your own `Bundle` with your
own packages of choice/implementations of choice.

`Novika.packages` is now `Bundle.available`