Errors: move details down, less whitespace
The latter could be considered a bad thing, but I'm tired of scrolling
thru all that whitespace... to find the error details at the very top.
This kind of fixes both. But now it feels too squeezed, and you can't
grey things out a bit because it's the terminal and you have, like, one
gray to choose from, one that might or might not end up looking good.