FFI: preallocate some arrays
Maybe improves performance, maybe does not. Probably doesn't, but still
a good practice -- after all, we *do* know the size of the arrays and
*do not* want the reallocs to trigger.

There isn't a lot of improvement because most function calls are small
in terms of argument count, and so, the initial capacity of Crystal's
Array is enough and no reallocs happen. Honestly, I have no idea.