Classifier: add number literals for bases 2, 8, 16
`0xbeef` etc. Add some tests too.

Does not seem to affect general performance much, even though I
resorted to regexes because I'm not in the mood of doing byte-
to-byte.

The thing does get a lot of hits because it's the last thing
before deciding an unclassified form is a word, and words are
probably the most commonly used form in Novika -- at slurp-time
for sure. So... Anyway. A closed issue is a good issue, I guess.

Closes #81.