Rewrite `entry:open` as a builtin for performance
Previously did two (!) lookups per call, which is hilariously stupid,
but this is a limitation of the `entry:` API.

How it is implemented still stinks, but at least it is faster.

Dumb stuff such as `100_000 compiledTimes: [ foo.x drop ]` where
`[ 100 $: x ] obj $: foo` is about 4x faster. When it comes to more
sophisticated code, e.g. tests, the difference is much less noticeable.