Fix `gsub`, introduce the chewbacca patch
Chewbacca chews away some word lookup by recursively replacing
everything from toplevel that resolves to a builtin with that builtin.

This is a fairly aggressive and expensive optimization, but it
is optimization nonetheless. Startup time degraded gravely, but
general runtime improved even in grim cases like the famous
`100_000 times: [ ]`.

More importantly, chewbacca is a self-optimization -- and therefore,
one of the features of Novika I'm glad to show off, after so much
work, even if it looks weird or makes the language load (but not work!)
somewhat slower.

And yes, I don't know why chewbacca.