Use `newBlockSet` in `gsub` and `flat:deep`
PRO: Makes a lot more sense to reuse what was written already

CON: F-cks up the relationship between the high-level and "lower-level"
parts of env/core even more. A lot of weird infinite cycles possible if
one is unaware which env/core/hl word is used where in env/core. In
other words, this is the most dangerous relationship: low-level stuff
talking to high-level stuff (the latter can sometimes be powered by the
very low-level stuff it's responding to!)