Fix recursive symlinks: ignore dir symlinks in `Disk#glob`
Now if you want to use a *directory* symlink, you have to refer to
it manually. File symlinks are still automatic. I have no idea who
is going to need (recursive) symlinks anyway, although they can provide
some indirection (and recursion in such case can be a mistake, and
mistakes clearly shouldn't blow up the whole program). To force the
use of manifests if you want symlinks, is a fair price to pay I think.

And the usual recursion depth limit to top it off, of course. I don't
want to spend another few weeks trying to implement a feature no one
is going to use.