Resolve env based on app's/lib's directory, not CWD
Also some very insignificant changes/refactors

Don't know how many bugs this introduces, but probably a lot.
The test suite is still very small to catch any of them, but
then the situations they're going to be encountered in are so
rare/useless it's weird to test them... Anyway, will see.