Keep a separate block store in `Dict` for predictable block keys
Set/get/has? etc. of blocks is still O(N). Equality when both sides
posess block keys is worse now, O(N^2) where N is the number of block
keys in both sides (it's the same; if the sizes differ we won't even
try, the dicts are not equal anyway).